- fix #includes (renamings) - removed now unused class XmlErrorHandler
git-svn-id: svn://svn.berlios.de/dsa-hl@3834 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/startup/src/RastullahMain.cpp
@@ -174,9 +174,9 @@ INT WINAPI WinMain( HINSTANCE hInst, HINSTANCE, LPSTR strCmdLine, INT )
     int argc;
     LPWSTR* argList = CommandLineToArgvW(GetCommandLineW(), &argc);
     char** argv = new char*[argc];
-    for (int argIdx = 0; argIdx < argc; argIdx++)
+    for (size_t argIdx = 0; argIdx < argc; argIdx++)
     {
-        int len = wcslen(argList[argIdx])+1;
+        size_t len = wcslen(argList[argIdx])+1;
         argv[argIdx] = new char[len];
         wcstombs(argv[argIdx], argList[argIdx], len);
         argv[argIdx][len-1] = '\0';
